+++ JestDotnet/XUnitTests/SettingsTests.cs
@@ -1,9 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Text.Encodings.Web;
-using System.Text.Json;
-using System.Text.Json.Serialization.Metadata;
-using System.Text.Unicode;
 using JestDotnet;
 using JestDotnet.Core.Settings;
 using Xunit;
@@ -79,62 +75,6 @@ public class SettingsTests
         finally
         {
             SnapshotSettings.NewLine = SnapshotSettings.DefaultNewLine;
-        }
-    }
-
-    [Fact]
-    public void ShouldSortAlphabeticallySimple()
-    {
-        SnapshotSettings.CreateSerializerOptions = () => new JsonSerializerOptions
-        {
-            WriteIndented = true,
-            Encoder = JavaScriptEncoder.Create(UnicodeRanges.All),
-            TypeInfoResolver = new DefaultJsonTypeInfoResolver
-            {
-                Modifiers = { AlphabeticalSortModifier.SortProperties }
-            },
-        };
-
-        try
-        {
-            var testObject = new Person
-            {
-                Age = 13,
-                DateOfBirth = new DateTime(2008, 7, 7),
-                FirstName = "John",
-                LastName = "Bam"
-            };
-
-            JestAssert.ShouldMatchSnapshot(testObject);
-        }
-        finally
-        {
-            SnapshotSettings.CreateSerializerOptions = SnapshotSettings.DefaultCreateSerializerOptions;
-        }
-    }
-
-    [Fact]
-    public void ShouldSortAlphabeticallyComplex()
-    {
-        SnapshotSettings.CreateSerializerOptions = () => new JsonSerializerOptions
-        {
-            WriteIndented = true,
-            Encoder = JavaScriptEncoder.Create(UnicodeRanges.All),
-            TypeInfoResolver = new DefaultJsonTypeInfoResolver
-            {
-                Modifiers = { AlphabeticalSortModifier.SortProperties }
-            },
-        };
-
-        try
-        {
-            var testObject = DataGenerator.GenerateComplexObjectData();
-
-            JestAssert.ShouldMatchSnapshot(testObject);
-        }
-        finally
-        {
-            SnapshotSettings.CreateSerializerOptions = SnapshotSettings.DefaultCreateSerializerOptions;
         }
     }